Get a set dataframe

This function queries the andes database and returns a dataframe containing the set data. TODO: The mission filter is not yet implemented everywhere, but should be added in the future. It currently uses the active mission

Usage

get_andes_set(andes_db_connection, mission_id)

Arguments

andes_db_connection

A connection object to the ANDES database.

mission_id

A string or integer representing the mission ID to filter the catches (default is NULL, which retrieves all catches).

Value

a dataframe containing andes catch rows

See also